requirements.txt
flask
pymongo
dnspython
gunicorn
python-dotenv
requests
flask_cors
flask_pymongo
python-json-logger